+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Help this project by [Donation](DONATE.md)
 Changes log
 -----------
 
+### 1.4.2
+
++ Added new attributes to `WHOIS` class: `emails`, `phone_numbers` and `fax_numbers`
++ Added a new property that contains expiration date(`expiration_date`) to `WHOIS` class
++ Improved WHOIS data parsing (For both human eyes and computer results)
+
 ### 1.4.1
 
 + Made whois run after initializing the `WHOIS` object optional.